requirements.txt
docopt
progress
GitPython>=2.1.0
uritemplate.py==2.0.0
github3.py==0.9.5
python-gitlab>=0.18
bitbucket-api
gogs-client>=1.0.3
